moved helper files from predict to dataset
